variable module created again
