+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Awesome Book</title>
-    <link rel="stylesheet" href="index.css" />
+    <link rel="stylesheet" href="style.css" />
   </head>
   <body>
     <header class="header">
@@ -51,5 +51,5 @@
     <p>All right reserved&copy;</p>
   </body>
   <script type="module" src="modules/dom.js"></script>
-  <script type="module" src="index.js"></script>
+  <script type="module" src="script.js"></script>
 </html>
