+++ modules/variable.js
@@ -7,7 +7,7 @@ export const addButton = document.querySelector('.add');
 export const addNew = document.getElementById('addNew');
 export const contact = document.getElementById('contact');
 export const list = document.getElementById('list');
- const containerAllbooks = document.querySelector('div.container');
+export const containerAllbooks = document.querySelector('div.container');
 export const inputForNewBook = document.querySelector('div.inputs');
 export const contactContainer = document.querySelector('div.contact');
-export let allBooks = [];
+export  let allBooks = [];
